memvid_api_server/requirements.txt
fastapi
uvicorn[standard]
gitpython
tree-sitter-language-pack
pathspec
memvid-sdk